+++ src/main/java/com/accenture/chickentest_app/ChickentestAppApplication.java
@@ -1,9 +1,11 @@
 package com.accenture.chickentest_app;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
 
 import java.sql.*;
 
@@ -11,6 +13,11 @@ import java.sql.*;
 @EnableConfigurationProperties
 @EntityScan(basePackages = {"com.accenture.chickentest_app.model"})
 public class ChickentestAppApplication {
+
+	@Bean
+	public ModelMapper modelMapper() {
+		return new ModelMapper();
+	}
 
 	public static void main(String[] args) {
 		SpringApplication.run(ChickentestAppApplication.class, args);