+++ src/main/java/com/accenture/chickentest_app/dto/ChickenDTO.java
@@ -1,0 +1,13 @@
+package com.accenture.chickentest_app.dto;
+
+import lombok.*;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+@Data
+public class ChickenDTO {
+    private double price;
+    private int daysToLive;
+}