+++ src/main/java/com/accenture/chickentest_app/model/Chicken.java
@@ -18,7 +18,22 @@ public class Chicken {
     Long id;
     @Column(name="price", nullable = false)
     double price;
-    // tentative
-    @Transient
-    int daysToLive;
+    // estos días son los que le quedan de vida, se modificarán
+    @Column(name="days_life", nullable = false)
+    int daysLived;
+    //estos días no se modifican, es el valor total de la vida
+    //para todas las gallinas será de 15
+    //se pide al crear una gallina
+    @Column(name="total_days", nullable = false)
+    final int totalDays = 15;
+
+    //tentative
+    //seteo de estado de gallina: lista a ser descartada
+    //debería ser guardado en la db?
+    //cuando se crea la gallina el estado es true
+    private boolean lifeStatus = true;
+
+    public boolean getLifeStatus() {
+        return lifeStatus;
+    }
 }