+++ src/main/java/com/accenture/chickentest_app/repository/EggRepository.java
@@ -7,4 +7,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface EggRepository extends JpaRepository<Egg, Long> {
+    Egg findAllById(Long l);
 }