+++ src/main/java/com/accenture/chickentest_app/service/ChickenService.java
@@ -1,7 +1,6 @@
 package com.accenture.chickentest_app.service;
 
 import com.accenture.chickentest_app.model.Chicken;
-import com.accenture.chickentest_app.model.Farmer;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,13 +8,17 @@ import java.util.Optional;
 
 @Service
 public interface ChickenService {
-    void addChicken(Chicken chicken);
+    Chicken addChicken(Chicken chicken);
 
     List<Chicken> getChickens();
 
     Optional<Chicken> getChicken(Long id);
 
-    void updateChicken(Long id, Chicken chicken);
+    Chicken updateChicken(Long id, Chicken chicken);
 
     void deleteChicken(Long id);
+    void addAllChicken(List<Chicken> chicken);
+    void advanceDays(int daysAdvanced);
+    void removeDead();
+    void dayChecker();
 }