+++ src/main/java/com/accenture/chickentest_app/service/FarmerService.java
@@ -1,5 +1,7 @@
 package com.accenture.chickentest_app.service;
 
+import com.accenture.chickentest_app.model.Chicken;
+import com.accenture.chickentest_app.model.Egg;
 import com.accenture.chickentest_app.model.Farmer;
 import org.springframework.stereotype.Service;
 
@@ -18,4 +20,9 @@ public interface FarmerService {
     void updateFarmer(Long id, Farmer farmer);
 
     void deleteFarmer(Long id);
+
+    boolean buyChicken(List<Chicken> chicken, Long farmerId);
+    boolean sellChicken(List<Long> chickenId, Long farmerId);
+    //void buyEgg(List<Egg> eggs);
+    //void sellEggs(List<int> eggsId);
 }