+++ src/main/java/com/accenture/chickentest_app/service/impl/FarmerServiceImpl.java
@@ -1,6 +1,9 @@
 package com.accenture.chickentest_app.service.impl;
 
+import com.accenture.chickentest_app.model.Chicken;
 import com.accenture.chickentest_app.model.Farmer;
+import com.accenture.chickentest_app.repository.ChickenRepository;
+import com.accenture.chickentest_app.repository.EggRepository;
 import com.accenture.chickentest_app.repository.FarmerRepository;
 import com.accenture.chickentest_app.service.FarmerService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +19,10 @@ public class FarmerServiceImpl implements FarmerService {
 
     @Autowired
     private FarmerRepository farmerRepository;
+    @Autowired
+    private ChickenRepository chickenRepository;
+    @Autowired
+    private EggRepository eggRepository;
 
     @Override
     public void addFarmer(Farmer farmer) {
@@ -51,5 +58,61 @@ public class FarmerServiceImpl implements FarmerService {
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Id inválido" + id));
 
         farmerRepository.delete(farmer);
+    }
+
+    @Override
+    public boolean buyChicken(List<Chicken> chicken, Long farmerId) {
+
+        double totalPrice = 0.0;
+        for(Chicken c : chicken) {
+            totalPrice += c.getPrice();
+        }
+
+        Farmer farmer = farmerRepository
+                .findById(farmerId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Id inválido" + farmerId));
+
+        if(farmer.getBalance() < totalPrice) {
+            return false;
+        } else {
+            //actualiza balance
+            farmer.setBalance(farmer.getBalance()-totalPrice);
+            //actualiza cantidad de gallinas
+            farmer.setChickenQuantity(chicken.size()+farmer.getChickenQuantity());
+            //actualiza cantidad de ganado total
+            farmer.setCattle(farmer.getCattle() + chicken.size());
+
+            chickenRepository.saveAll(chicken);
+            return true;
+        }
+    }
+
+    @Override
+    public boolean sellChicken(List<Long> chickenId, Long farmerId) {
+
+        double totalPrice = 0.0;
+        for(Long id : chickenId) {
+            totalPrice += chickenRepository.getById(id).getPrice();
+        }
+
+        Farmer farmer = farmerRepository
+                .findById(farmerId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Id inválido" + farmerId));
+
+        //chequea si se tiene suficiente ganado para vender
+        if(farmer.getCattle() < farmer.getMinFarmQuantity()) {
+            //actualiza balance
+            farmer.setBalance(farmer.getBalance()+totalPrice);
+            //actualiza cantidad de gallinas
+            farmer.setChickenQuantity(chickenId.size()-farmer.getChickenQuantity());
+            //actualiza cantidad ganado total
+            farmer.setCattle(farmer.getCattle() - chickenId.size());
+            //elimina las gallinas vendidas
+            chickenRepository.deleteAllById(chickenId);
+
+            return true;
+        } else {
+            return false;
+        }
     }
 }