+++ src/main/resources/CHANGELOG.md
@@ -7,3 +7,18 @@ _TO-DO_
 - make service method that looks for chicken and egg that are
   ready to transform/die and include them in the report (returning
   the id for them)
+
+-------------------------------------
+### v0.0.1 (Nov 19th, 2024)
+
+_TO-DO_
+- make service method that looks for chicken and egg that are
+  ready to transform/die and include them in the report (returning
+  the id for them)
+
+_DONE_
+  * chickenDTO 
+  * sell and buy methods and endpoints for Chicken
+  * method to advance days for Chicken
+  * method to remove dead Chicken from repository
+  * method to set status as false for death